% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installSpaDES}
\alias{installSpaDES}
\title{Install SpaDES packages, making sure to \code{update.packages} first}
\usage{
installSpaDES(
  ask = FALSE,
  type,
  libPath = .libPaths()[1],
  versions = c(SpaDES.core = "1.0.5", SpaDES.tools = "0.3.6"),
  dontUpdate = c("scam")
)
}
\arguments{
\item{ask}{Passed to \code{update.packages}.}

\item{type}{passed to both \code{update.packages} and \code{install.packages}.
This will set \code{"binary"} on Windows, if not set, to get the binary packages from CRAN.}

\item{libPath}{Passed to both \code{update.packages(lib.lob = libPath)} and
\code{install.packages(lib = libPath, ...)}}

\item{versions}{named list of package versions to install}

\item{dontUpdate}{character vector of packages not to update}
}
\description{
Install SpaDES packages, making sure to \code{update.packages} first
}
