% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packagesInModules}
\alias{packagesInModules}
\alias{metadataInModules}
\title{Extract element from SpaDES module metadata}
\usage{
packagesInModules(modules, modulePath = getOption("spades.modulePath"))

metadataInModules(
  modules,
  metadataItem = "reqdPkgs",
  modulePath = getOption("spades.modulePath"),
  needUnlist
)
}
\value{
A character vector of sorted, uniqued packages that are identified in all named
modules, or if \code{modules} is omitted, then all modules in \code{modulePath}.
}
\description{
Parses module code, looking for the \code{metadataItem} (default = \code{"reqdPkgs"})
element in the \code{defineModule} function.
}
