% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packagesInModules}
\alias{packagesInModules}
\title{Extract reqdPkgs from SpaDES modules}
\usage{
packagesInModules(modulePath = options("spades.modulePath"), modules)
}
\value{
A character vector of sorted, uniqued packages that are identified in all named
modules, or if \code{modules} is omitted, then all modules in \code{modulePath}.
}
\description{
Parses module code, looking for the \code{reqdPkgs} element in the \code{defineModule} function.
}
