% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsSetup.R
\name{setLinuxBinaryRepo}
\alias{setLinuxBinaryRepo}
\title{Setup for binary linux repositories}
\usage{
setLinuxBinaryRepo(
  CRAN = "https://cloud.r-project.org",
  binaryLinux = "https://packagemanager.rstudio.com/"
)
}
\arguments{
\item{CRAN}{The default CRAN repository to use, when not using a binary Linux repo}

\item{binaryLinux}{A binary linux repository.}
}
\description{
This function is run for its side effect on the options("repos") and the options("HTTPUserAgen")
}
