% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{sourcePkgs}
\alias{sourcePkgs}
\title{Character vector of packages that must be installed from source in Linux-alikes}
\usage{
sourcePkgs(additional = "")
}
\arguments{
\item{additional}{Character string indicating additional packages to add to the
default set (to see default set, run \code{sourcePkgs()}).}
}
\description{
Character vector of packages that must be installed from source in Linux-alikes
}
