% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{DESCRIPTIONFileDeps}
\alias{DESCRIPTIONFileDeps}
\alias{DESCRIPTIONFileVersionV}
\alias{getGitHubFile}
\alias{isWindows}
\alias{messageDF}
\title{non-exported objects and functions from other packages}
\usage{
DESCRIPTIONFileDeps(
  desc_path,
  which = c("Depends", "Imports", "LinkingTo"),
  keepVersionNumber = TRUE,
  purge = getOption("Require.purge", FALSE),
  keepSeparate = FALSE
)

DESCRIPTIONFileVersionV(file, purge = getOption("Require.purge", FALSE))

getGitHubFile(
  pkg,
  filename = "DESCRIPTION",
  purge = getOption("Require.purge", FALSE)
)

isWindows()

messageDF(df, round)
}
\description{
non-exported objects and functions from other packages
}
\keyword{internal}
