% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installSourcePackages}
\alias{installSourcePackages}
\title{Install spatial packages}
\usage{
installSourcePackages(
  pkgs = c("rgeos", "rgdal", "terra", "sf", "sp", "raster", "igraph", "units", "qs",
    "Rcpp", "RcppParallel", "cpp11"),
  libPath = .libPaths()[1],
  repos = "https://cloud.r-project.org",
  forceWindows = FALSE
)
}
\arguments{
\item{repos}{URL of CRAN mirror to use to fetch source packages}
}
\description{
(Re)install spatial packages that require GDAL/GEOS/PROJ, from source to ensure they are properly
liked to these external libraries.
}
\note{
if installing on macOS, homebrew installation of GDAL etc. is required.
}
