% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installSpaDES}
\alias{installSpaDES}
\title{Install SpaDES packages, making sure to \code{update.packages} first}
\usage{
installSpaDES(
  ask = FALSE,
  type,
  libPath = .libPaths()[1],
  fromSource = c("igraph", "rgeos", "rgdal", "terra", "sf", "units", "qs", "sp",
    "Rcpp", "RcppParallel", "cpp11"),
  versions = c("SpaDES.core (>=1.0.9)", "SpaDES.tools (>= 0.3.9)"),
  dontUpdate = c("scam"),
  SpaDES.project = TRUE
)
}
\arguments{
\item{ask}{Passed to \code{update.packages}.}

\item{type}{passed to both \code{update.packages} and \code{install.packages}.
This will be set to \code{"binary"} on Windows, if not set, to get the binary packages from CRAN.}

\item{libPath}{Passed to both \code{update.packages(lib.lob = libPath)} and
\code{install.packages(lib = libPath, ...)}}

\item{fromSource}{A character vector of packages that must be installed from
source on Linux-alikes, even if the \code{options("repos")} is pointing to a binary
repository of binary packages. The default is all the spatial packages, plus
a few others. Has no effect if}

\item{versions}{named list of package versions to install}

\item{dontUpdate}{character vector of packages not to update}

\item{SpaDES.project}{logical. If TRUE, the default, then the SpaDES.project will
also be installed. This is not on CRAN. It will first attempt to install from
predictiveecology.r-universe.dev. If that fails, then it will install from source
from github.com/PredictiveEcology/SpaDES.project}
}
\description{
Install SpaDES packages, making sure to \code{update.packages} first
}
