% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installSpatialPackages}
\alias{installSpatialPackages}
\title{Install spatial packages}
\usage{
installSpatialPackages(repos = "https://cran.rstudio.com")
}
\arguments{
\item{repos}{URL of CRAN mirror to use to fetch source packages}
}
\description{
(Re)install spatial packages that require GDAL/GEOS/PROJ, from source to ensure they are properly
liked to these external libraries.
}
\note{
if installing on macOS, homebrew installation of GDAL etc. is required.
}
