% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installSpatialPackages}
\alias{installSpatialPackages}
\title{Install spatial packages}
\usage{
installSpatialPackages(
  pkgs = c("rgeos", "sp", "raster", "terra", "lwgeom", "sf", "rgdal"),
  repos = "https://cloud.r-project.org",
  libPath = .libPaths()[1],
  forceSourceOnWindows = FALSE
)
}
\arguments{
\item{pkgs}{Character vector of package names to install using \code{type = "source"}}

\item{repos}{URL of CRAN mirror to use to fetch source packages}

\item{libPath}{Passed to both \code{update.packages(lib.lob = libPath)} and
\code{install.packages(lib = libPath, ...)}}

\item{forceWindows}{Logical. If \code{TRUE}, then this will install from source on Windows,
which is often unnecessary for e.g., spatial packages.}
}
\description{
(Re)install spatial packages that require GDAL/GEOS/PROJ, from source to ensure they are properly
liked to these external libraries.
}
\note{
if installing on macOS, homebrew installation of GDAL etc. is required.
}
