% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionsSetup.R
\name{setLinuxBinaryRepo}
\alias{setLinuxBinaryRepo}
\title{Setup for binary Linux repositories}
\usage{
setLinuxBinaryRepo(
  CRAN = getOption("repos", "https://cloud.r-project.org/"),
  binaryLinux = "https://packagemanager.rstudio.com/"
)
}
\arguments{
\item{CRAN}{The default CRAN repository to use, when not using a binary Linux repo.}

\item{binaryLinux}{A CRAN repository serving binary Linux packages.}
}
\description{
Enable use of binary package builds for Linux from the Rstudio Package Manager repo.
This will set the \code{repos} option, affecting the current R session.
}
