% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-install-package.R
\docType{package}
\name{SpaDES.install-package}
\alias{SpaDES.install}
\alias{SpaDES.install-package}
\title{Helpers for installing \code{SpaDES} package and module dependencies}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spades-install.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.install}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.install/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alex.chubaty@gmail.com} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
